Create mapsizecalc
This file is a javascript macro for Foundry VTT.  This assists the GM with determining what the map size should be based on the background uploaded and determining how many pixels on that image would constitute a size 0 character pog.  This sets the recommended map size for unit of measure 1" with 50x50 gridless or grid maps.